# chapter3-mygram/docs/swagger.yaml
basePath: /
definitions:
...
  model.ErrorResponse:
    properties:
      code:
        type: integer
      errors: {}
      status:
        type: string
    type: object
...
  model.SocialCreateReq:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    required:
      - name
      - social_media_url
    type: object
  model.SocialDeleteRes:
    properties:
      id:
        type: string
    type: object
  model.SocialResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.SocialUpdateReq:
    properties:
      name:
        type: string
      social_media_url:
        type: string
    required:
      - name
      - social_media_url
    type: object
  model.SocialUpdateRes:
    properties:
      id:
        type: string
    type: object
  model.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.UserLoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
      - password
      - username
    type: object
  model.UserLoginRes:
    properties:
      token:
        type: string
    type: object
  model.UserRegisterReq:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
      - age
      - email
      - password
      - username
    type: object
  model.UserRegisterRes:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      password:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Mygram API untuk last project dari DTS dan Hacktiv8.
  title: Hacktiv8 Mygram API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
        - application/json
      description: login user using username and password
      parameters:
        - description: request is required
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/model.UserLoginReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    $ref: "#/definitions/model.UserLoginRes"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      summary: login user
      tags:
        - User
...
  /register:
    post:
      consumes:
        - application/json
      description: filled some form for registration
      parameters:
        - description: request is required
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/model.UserRegisterReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    $ref: "#/definitions/model.UserRegisterRes"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      summary: register user
      tags:
        - User
  /social-media:
    get:
      consumes:
        - application/json
      description: get all social media
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/model.SocialResponse"
                    type: array
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: get all social media
      tags:
        - Social Media
    post:
      consumes:
        - application/json
      description: create social media for a particular user
      parameters:
        - description: request is required
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/model.SocialCreateReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    $ref: "#/definitions/model.SocialResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: create social media
      tags:
        - Social Media
  /social-media/{social_media_id}:
    delete:
      consumes:
        - application/json
      description: delete social media
      parameters:
        - description: social_media_id
          in: path
          name: social_media_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    $ref: "#/definitions/model.SocialDeleteRes"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: delete social media
      tags:
        - Social Media
    get:
      consumes:
        - application/json
      description: get one social media
      parameters:
        - description: social_media_id
          in: path
          name: social_media_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    $ref: "#/definitions/model.SocialResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: get one social media
      tags:
        - Social Media
    put:
      consumes:
        - application/json
      description: update social media
      parameters:
        - description: social_media_id
          in: path
          name: social_media_id
          required: true
          type: string
        - description: request is required
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/model.SocialUpdateReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/model.SuccessResponse"
              - properties:
                  data:
                    $ref: "#/definitions/model.SocialUpdateRes"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
              - $ref: "#/definitions/model.ErrorResponse"
              - properties:
                  errors:
                    type: object
                type: object
      security:
        - BearerAuth: []
      summary: update social media
      tags:
        - Social Media
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"